Figured out Trapezoid overshoot issue
If kS, kV, or kA are too high, the profile will overshoot and not stop overshooting. Split driveStraightTrapezoid profile into 2 profiles (one that regenerates the profile every cycle and one that sticks with only one profile). Next step is to add PID to both and determine which we want to keep/continue on with for actual turning. Still working in linear for now.